feat(#15): add auth mutations and queries
