+++ src/server/auth/mutations/index.ts
@@ -1,0 +1,2 @@
+export { useLogin } from './useLogin';
+export { useLogout } from './useLogout';
