+++ src/server/auth/mutations/useLogin.ts
@@ -1,6 +1,6 @@
 import type { Provider } from '@supabase/supabase-js';
 import { useMutation } from '@tanstack/react-query';
-import { supabase } from '../provider';
+import { supabase } from '../../provider';
 
 const loginGithub = async (provider: Provider) => {
   const { data } = await supabase.auth.signInWithOAuth({
