+++ src/server/auth/queries/index.ts
@@ -1,0 +1,1 @@
+export { useUserSession } from './useUserSession';
