+++ src/server/auth/queries/keys.ts
@@ -1,0 +1,3 @@
+export const AUTH_KEYS = {
+  userSession: ['userSession'],
+} as const;
