--- src/server/auth/useLogout.ts
@@ -1,12 +1,0 @@
-import { useMutation } from '@tanstack/react-query';
-import { supabase } from '../provider';
-
-const logout = async () => {
-  await supabase.auth.signOut();
-};
-
-export const useLogout = () => {
-  return useMutation({
-    mutationFn: logout,
-  });
-};
